TASK3 - Realizando Ajustes - Move automapper para pasta própria - Adiciona novo pacote para injeção de dependência - Ajusta interface mapper
