+++ src/Backend/BrazilianAddresses.Application/Services/Automapper/AutomapperConfig.cs
@@ -2,7 +2,7 @@
 using BrazilianAddresses.Communication.Requests;
 using BrazilianAddresses.Domain.Entities;
 
-namespace BrazilianAddresses.Application.Services
+namespace BrazilianAddresses.Application.Services.Automapper
 {
     public class AutomapperConfig : Profile
     {
